requirements.txt
colorama==0.4.6
customtkinter==5.2.2
discord.py==2.4.0
protobuf==5.28.2
google.generativeai
